backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.13.4
billiard==4.2.2
bleach==6.2.0
cachetools==6.2.0
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect==1.3.0
cssutils==2.11.1
Django==5.2.3
django-admin-sortable2==2.2.8
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
idna==3.10
kombu==5.5.4
lxml==6.0.2
more-itertools==10.8.0
packaging==25.0
pillow==11.2.1
premailer==3.10.0
prompt_toolkit==3.0.52
PyMuPDF==1.26.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
mysqlclient==2.2.0
redis==6.4.0
requests==2.32.5
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
text-unidecode==1.3
tinycss2==1.4.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
